requirements.txt
aiohttp
basc_py4chan
nest_asyncio
nudenet
openai==0.28
Pillow
pyperclip
python-telegram-bot
Requests
tensorflow-cpu
aiogram
python-dotenv
fastapi
uvicorn
honcho